requirements.txt
django
djangorestframework
python-dotenv
bleach
redis
gunicorn
dj-database-url
psycopg2-binary
whitenoise
